+++ spiketools/plts/cell.py
